TwilioService: Allow skipping of Twilio client
Added a new environment variable to activate or deactive Twilio. This is so I don't use up all of my Twilio credits but can turn it on if I need to test it. 

It is automatically turned on in production.

Also, I was missing a 'to' variable earlier and changed my hardcoded phone number to this new variable