ScoreController: Ability to store scores
Added the ability to store scores. 

It first checks that the user exists with that phone number. If not, it returns a 404 error. I decided against having Twilio send a text because if they aren't in the database, they shouldn't be using up my text credits anyway. 

If the user does exist, we count all of the scores for today. I have no clue how timezones are factored in here. If the user has submitted a score for today, the user is blocked from submitting a second one.

Added a new api route for storing scores.