ScoreController: Account for 'X' case
Originally, the way to decide the score was to count the lines. This was foolish, as it didn't account for the case where the user did not get the wordle that day. So, we now split the string and look for the X/6 portion and then the score is the number before the slash. And if it an "X", the score is saved as 0.