TwilioService: sendScoreBoardMessage
ScoreController now stores "X" scores as the value of 7 and not 0. This is for sorting purposes.

Added new method: sendScoreBoardMessage which takes in a phone number of who it is sending the message to. 

The method first retrieves all of the scores and orders them best to worst (lowest to highest).

Then, it creates an array of the rankings. It goes through all of the scores and accounts for ties. It then adds the place (e.g. 1st), name, and score (e.g. '1/6'). It also checks if the '7' and if so, changes it to be an 'X' for 'X/6'.

I also added a new method that I got online which adds the ordinal number suffix. I putting my full faith in the random coder online. Thanks.